ems_backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
coverage==6.4.1
cryptography==37.0.2
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.7.1
idna==3.3
pipreqs==0.4.11
platformdirs==2.5.2
pycparser==2.21
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.15.0
yarg==0.1.9
